dialer: move former dialerapi inside internal
Now everything inside dialer more or less makes sense and we
have got rid of a good amount of spaghetti already.
